create index.js in the folder controllers/contacts;  shared func
