+++ controllers/contacts/deleteContactById.js
@@ -1,0 +1,16 @@
+const {Contact} = require("../../models/contact")
+const { HttpError, ctrlWrapper } = require("../../helpers");
+
+const deleteContactById = async (req, res) => {
+    const { contactId } = req.params;
+    const result = await Contact.findByIdAndRemove(contactId);
+    if (!result) {
+      throw HttpError(404, "Not found");
+    }
+    res.json({
+      message: "contact delete",
+    });
+  };
+  module.exports = {
+    deleteContactById: ctrlWrapper(deleteContactById),
+  }
