+++ controllers/contacts/index.js
@@ -1,0 +1,16 @@
+const { getAllContacts } = require("./getAllContacts");
+const { getById }= require('./getById');
+const { addContact } = require("./addContact");
+const { updateContactById } = require("./updateContactById");
+const { updateFavorite } = require("./updateFavorite");
+const { deleteContactById } = require("./deleteContactById");
+
+module.exports = {
+    getAllContacts,
+    getById,
+    addContact,
+    updateContactById,
+    updateFavorite,
+    deleteContactById,
+
+};
