+++ models/index.js
@@ -1,0 +1,7 @@
+const { Contact } = require("./contact");
+const { schemas }= require("./contact");
+
+module.exports = {
+    Contact,
+    schemas,
+}
