+++ routes/api/contacts.js
@@ -1,12 +1,12 @@
 const express = require("express");
+const router = express.Router();
 
-const ctrl = require("../../controllers/contacts/contacts");
+const ctrl = require("../../controllers/contacts");
 
 const {validateBody} = require("../../middlewares");
 const {isValidId}= require("../../middlewares");
 const { schemas }= require("../../models/contact");
 
-const router = express.Router();
 
 router.get("/", ctrl.getAllContacts);
